backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pandas==2.1.3
numpy==1.26.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
